requirements.txt
flask
pdfplumber
tabulate
gspread
google-auth
gunicorn